\name{ALPODSclassifier}
\alias{ALPODSclassifier}
\alias{ALPODSclassifierV2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ALPODSclassifier
}
\description{
rule based classifier and subpopulations corresponding to RuleCls generated by ALPODS described in [Ultsch et al, 2022].
}
\usage{
ALPODSclassifier(Data, Header,Type="V2")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{
[1:d,1:n]        data array of d cases with n variables
}
  \item{Header}{
[1:n]        character vector of variable names
}
\item{Type}{
  Two possible choices: the newer "V2" is used crrently per default, "V1" remains for backcompatiblity reasons
}
}
\details{
The classifier is automaticall generated by the ALPODS XAI. Historically there are two versions, the newer version V2 is used currently but the older version is still attached for reproducibility of results.
}
\value{
List of
\item{RuleCls }{[1:d)             vector of classes, d integer numbers, number k indicates class k}
\item{PopulationNumber }{ vector of classes, d integer numbers, number i indicates subpopulation}
}
\references{
[Ultsch et al., 2022]  Ultsch, A., Hoffman, J., Röhnert, M., Von Bonin, M., Oelschlägel, U., Brendel, C., & Thrun, M. C.: An explainable AI system for the diagnosis of high-dimensional biomedical data, in revision, 2022.
}
\author{
XAI-ALPODS
}
\note{
 V 1.4 automatically generated by XAI ALPODS in matlab [Ultsch et al., 2021] 
}

\examples{
data("PreprocessedDataSample")
Data=TransV[[1]]
model=ALPODSclassifier(Data,colnames(Data))
}
