\name{GeneratePopulations}
\alias{GeneratePopulations}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GeneratePopulations
}
\description{
Generate Populations based on data and given ALPODS classifier defined in \code{\link{ALPODSclassifier}}.
}
\usage{
GeneratePopulations(DataV, FUN, cl)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DataV}{
[1:m] list if numerical data matrices [1:n?,1:d?] of n cases and d features, n and d can vary in each list element
}
  \item{FUN}{
function of \code{\link{ALPODSclassifier}}
}
  \item{cl}{
if missing, generates one core less than exisits on pc and computes on each core the populations
if \code{NULL} uses one core
otherwise specify manually using cl=parallel::makeCluster(...)
}
}
\details{
Population freuencies are generated based on a classifier generated by the ALPODS XAI. It is expected that the data is normalized using \code{\link{PreprocessedDataSample}}.
}
\value{
List of
\item{PopsMat}{numerical matrix [1:m,1:NoPopulations] of population frequencies based on the classifier}
\item{Cls}{[1:NoPopulations] numerical vector of classification of weak supervised ML}
\item{OutputFUN}{Full output of function given in \code{FUN}}
}

\author{
Michael Thrun
}

\seealso{
\code{\link{ALPODSclassifier}}, \code{\link{PreprocessedDataSample}}
}
\examples{
data("PreprocessedDataSample")
ListeV=GeneratePopulations(TransV,"ALPODSclassifier",NULL)
round(ListeV$PopsMat*100,3) #percentages

ListeV=GeneratePopulations(TransV,"ALPODSclassifierV2",NULL)
round(ListeV$PopsMat*100,3) #percentages
}
