\name{MTcinderellaALPODSClassify}
\alias{MTcinderellaALPODSClassify}

\title{
MTcinderellaALPODSClassify is an internal function
}
\description{
function ensures back compatibility with performed analysis 
}
\usage{
MTcinderellaALPODSClassify(Data, Header)
}

\arguments{
  \item{Data}{
(1:d,1:n)        data array of d cases with n variables
}
  \item{Header}{
(1:n,)        array of variable names
}
}
\details{
please ignore this function and see, code{\link{ALPODSclassifier}}
}
\value{
List of
\item{RuleCls }{[1:d)             vector of classes, d integer numbers, number k indicates class k}
\item{PopulationNumber }{ vector of classes, d integer numbers, number i indicates subpopulation}
}
\author{
XAI-ALPODS
}
\note{
 V 2.0 automatically generated by XAI ALPODS in R
}


\seealso{
code{\link{ALPODSclassifier}}
}

