\name{PreprocessedDataSample}
\alias{PreprocessedDataSample}
\docType{data}
\title{
PreprocessedDataSample
}
\description{
samples of log transformed h1 panel data of navios
}
\usage{data("PreprocessedDataSample")}
\format{
\code{TransV} is list of two numerical matrizes of data
}
\details{
small preprocessed annonymized sample of data
}
\source{
Dr. Joerg Hoffmann
}

\examples{
data(PreprocessedDataSample)
#str(TransV)
}
\keyword{datasets}
